move ModelManager initialization into its own module and restore embedding support
